# fix Base.Fix1/2 error