--- a/test/references/ref20_2_sha1.txt
+++ b/test/references/ref20_2_sha1.txt
@@ -1,0 +1,1 @@
+ff518b2235f97cdcf992bdb862a87cce9ee3b841