… references/ref20_1_crc.txt
… test/references/ref20_1_crc.txt
… test/references/ref20_1_sha1.txt
… test/references/ref20_1_sha256.txt
… test/references/ref20_2_crc32c.txt
… test/references/ref20_2_sha256.txt